+++ src/Processors/QueryPlan/Optimizations/considerEnablingParallelReplicas.cpp
@@ -1,0 +1,376 @@
+#include <Processors/QueryPlan/Optimizations/considerEnablingParallelReplicas.h>
+
+#include <Interpreters/PreparedSets.h>
+#include <Processors/QueryPlan/CreatingSetsStep.h>
+#include <Processors/QueryPlan/ExpressionStep.h>
+#include <Processors/QueryPlan/FilterStep.h>
+#include <Processors/QueryPlan/JoinLazyColumnsStep.h>
+#include <Processors/QueryPlan/ReadFromMergeTree.h>
+#include <Processors/QueryPlan/ReadFromRemote.h>
+#include <Processors/QueryPlan/UnionStep.h>
+#include <Processors/QueryPlan/Optimizations/Optimizations.h>
+#include <Processors/QueryPlan/Optimizations/RuntimeDataflowStatistics.h>
+#include <Processors/QueryPlan/Optimizations/Utils.h>
+#include <Common/Exception.h>
+#include <Common/Logger.h>
+#include <Common/logger_useful.h>
+#include <Common/typeid_cast.h>
+
+#include <map>
+
+namespace DB
+{
+
+namespace ErrorCodes
+{
+extern const int LOGICAL_ERROR;
+}
+
+namespace QueryPlanOptimizations
+{
+
+/// Find the top node of the parallel replicas plan. E.g.:
+///
+/// Expression ((Project names + Projection))
+///  MergingAggregated
+///    Union
+///      Aggregating  <-- this node is the last plan step to be executed on replicas
+///        Expression (Before GROUP BY)
+///          Expression ((WHERE + Change column names to column identifiers))
+///            ReadFromMergeTree (default.hits)
+///      ReadFromRemoteParallelReplicas (Query: ... Replicas: ...)
+///
+static QueryPlan::Node * findTopNodeOfReplicasPlan(QueryPlan::Node * plan_with_parallel_replicas_root)
+{
+    QueryPlan::Node * replicas_plan_top_node = nullptr;
+
+    Stack stack;
+    stack.push_back({.node = plan_with_parallel_replicas_root});
+
+    while (!stack.empty())
+    {
+        auto & frame = stack.back();
+
+        /// Currently the approach is very simple: we look for Union step in the plan tree,
+        /// and consider its children. The first child that is not ReadFromParallelRemoteReplicas
+        /// is considered the top node of replicas plan.
+        if (typeid_cast<UnionStep *>(frame.node->step.get()))
+        {
+            bool found_read_from_parallel_replicas = false;
+
+            for (const auto & child : frame.node->children)
+            {
+                auto * node = child;
+                /// ExpressionStep can be placed on top of ReadFromRemoteParallelReplicas
+                if (typeid_cast<const ExpressionStep *>(node->step.get()) || typeid_cast<const FilterStep *>(node->step.get()))
+                {
+                    chassert(!node->children.empty());
+                    node = node->children.front();
+                }
+                if (typeid_cast<const DelayedCreatingSetsStep *>(node->step.get())
+                    || typeid_cast<const CreatingSetsStep *>(node->step.get()))
+                {
+                    chassert(!node->children.empty());
+                    node = node->children.front();
+                }
+                if (!typeid_cast<const ReadFromParallelRemoteReplicasStep *>(node->step.get()))
+                {
+                    if (replicas_plan_top_node)
+                    {
+                        // TODO(nickitat): support multiple read steps with parallel replicas
+                        LOG_DEBUG(getLogger("optimizeTree"), "Top node for parallel replicas plan is already found");
+                        return nullptr;
+                    }
+
+                    replicas_plan_top_node = node;
+                }
+                else
+                {
+                    found_read_from_parallel_replicas = true;
+                }
+            }
+
+            /// We found pattern
+            ///     Union
+            ///       ReadFromParallelRemoteReplicas
+            ///       <replicas_plan_top_node>
+            if (replicas_plan_top_node && found_read_from_parallel_replicas)
+                break;
+        }
+
+        /// Traverse all children first.
+        if (frame.next_child < frame.node->children.size())
+        {
+            auto next_frame = Frame{.node = frame.node->children[frame.next_child]};
+            ++frame.next_child;
+            stack.push_back(next_frame);
+            continue;
+        }
+
+        stack.pop_back();
+    }
+
+    return replicas_plan_top_node;
+}
+
+/// Now when we found the top node of replicas plan, we need to find the corresponding node in the single node plan.
+/// The working principle behind automatic parallel replicas is that we use statistics collected during execution of single-node plan
+/// to estimate whether parallel replicas will be beneficial for the query or not. For that, we need to estimate how much data
+/// replicas will send to the initiator. To do that, we found the node that will be at the top of replicas plan (e.g. Aggregating step in the example above),
+/// and ask it collect statistics on the number of bytes it'd send to the initiator if we executed the query with parallel replicas.
+static std::pair<const QueryPlan::Node *, size_t> findCorrespondingNodeInSingleNodePlan(
+    const QueryPlan::Node & final_node_in_replica_plan,
+    QueryPlan::Node & parallel_replicas_plan_root,
+    QueryPlan::Node & single_replica_plan_root)
+{
+    auto pr_node_hashes = calculateHashTableCacheKeys(parallel_replicas_plan_root);
+    if (auto it = pr_node_hashes.find(&final_node_in_replica_plan); it != pr_node_hashes.end())
+    {
+        auto nopr_node_hashes = calculateHashTableCacheKeys(single_replica_plan_root);
+
+        for (const auto & [nopr_node, nopr_hash] : nopr_node_hashes)
+        {
+            if (nopr_hash == it->second)
+            {
+                if (!nopr_node->step->supportsDataflowStatisticsCollection())
+                {
+                    LOG_DEBUG(
+                        getLogger("optimizeTree"),
+                        "Step ({}) doesn't support dataflow statistics collection. Skipping statistics collection",
+                        nopr_node->step->getName());
+                    return std::make_pair(nullptr, 0);
+                }
+
+                LOG_DEBUG(getLogger("optimizeTree"), "Found matching node in original plan: {}", nopr_node->step->getName());
+                return std::make_pair(nopr_node, nopr_hash);
+            }
+        }
+        LOG_DEBUG(getLogger("optimizeTree"), "Cannot find step with matching hash in single-node plan");
+        return std::make_pair(nullptr, 0);
+    }
+    else
+    {
+        throw Exception(ErrorCodes::LOGICAL_ERROR, "Cannot find replicas_plan_top_node in hash table");
+    }
+}
+
+static ReadFromMergeTree * findReadingStep(const QueryPlan::Node & top_of_single_replica_plan)
+{
+    const auto * reading_step = &top_of_single_replica_plan;
+    while (reading_step && !reading_step->children.empty())
+    {
+        // TODO(nickitat): support multiple read steps with parallel replicas
+        const auto * lazy_joining = typeid_cast<const JoinLazyColumnsStep *>(reading_step->step.get());
+
+        if (!lazy_joining && reading_step->children.size() > 1)
+            return nullptr;
+        reading_step = reading_step->children.front();
+    }
+
+    chassert(reading_step);
+    if (auto * read_from_merge_tree = typeid_cast<ReadFromMergeTree *>(reading_step->step.get()))
+        return read_from_merge_tree;
+
+    LOG_DEBUG(
+        getLogger("optimizeTree"),
+        "Cannot find ReadFromMergeTree step in single-replica plan (found {}). Skipping optimization",
+        reading_step->step->getName());
+    return nullptr;
+}
+
+/// Transplant the sets from the single-replica plan to the parallel-replicas plan once we decided to enable parallel replicas
+static void moveSetsFromLocalPlanToReplicasPlan(const QueryPlan & single_replica_plan, const QueryPlan & parallel_replicas_plan)
+{
+    Stack stack;
+    std::map<FutureSet::Hash, SetAndKeyPtr> sets_map;
+
+    // Create a map: set_key -> set
+    stack.clear();
+    traverseQueryPlan(
+        stack,
+        *single_replica_plan.getRootNode(),
+        [&](auto & frame_node)
+        {
+            if (auto * creating_sets_step = typeid_cast<DelayedCreatingSetsStep *>(frame_node.step.get()))
+            {
+                const auto sets = creating_sets_step->detachSets();
+                for (const auto & future_set : sets)
+                {
+                    if (auto set = future_set->detachSetAndKey())
+                        sets_map[future_set->getHash()] = std::move(set);
+                }
+            }
+        });
+
+    // Now transplant the sets
+    stack.clear();
+    traverseQueryPlan(
+        stack,
+        *parallel_replicas_plan.getRootNode(),
+        [&](auto & frame_node)
+        {
+            if (const auto * creating_sets_step = typeid_cast<DelayedCreatingSetsStep *>(frame_node.step.get()))
+            {
+                for (const auto & future_set : creating_sets_step->getSets())
+                {
+                    if (auto it = sets_map.find(future_set->getHash()); it != sets_map.end())
+                    {
+                        future_set->replaceSetAndKey(it->second);
+                    }
+                    else
+                    {
+                        throw Exception(
+                            ErrorCodes::LOGICAL_ERROR, "Cannot find a matching set in the map of sets from single-replica plan");
+                    }
+                }
+            }
+        });
+}
+
+void considerEnablingParallelReplicas(
+    const QueryPlanOptimizationSettings & optimization_settings, QueryPlan::Node & root, QueryPlan & query_plan)
+{
+    if (!optimization_settings.automatic_parallel_replicas_mode
+        || !optimization_settings.query_plan_with_parallel_replicas_builder
+        || optimization_settings.parallel_replicas_enabled)
+        return;
+
+    // Cannot guarantee projection usage with parallel replicas
+    if (optimization_settings.force_use_projection)
+        return;
+
+    Stack stack;
+    // Technically, it isn't required for all steps to support dataflow statistics collection,
+    // but only for those that we will actually instrument (see `setRuntimeDataflowStatisticsCacheUpdater` calls below).
+    // However, currently only relatively simple plans are supported (no JOINs, CreatingSets from subqueries, UNIONs, etc.),
+    // since all these steps obviously don't support statistics collection, `supportsDataflowStatisticsCollection` is handy to check if the plan is simple enough.
+    bool plan_is_simple_enough = true;
+    traverseQueryPlan(
+        stack,
+        root,
+        [&](auto & frame_node)
+        {
+            plan_is_simple_enough &= frame_node.step->supportsDataflowStatisticsCollection()
+                || typeid_cast<const DelayedCreatingSetsStep *>(frame_node.step.get())
+                || typeid_cast<const CreatingSetsStep *>(frame_node.step.get());
+        });
+    if (!plan_is_simple_enough)
+    {
+        LOG_DEBUG(getLogger("optimizeTree"), "Some steps in the plan don't support dataflow statistics collection. Skipping optimization");
+        return;
+    }
+
+    auto plan_with_parallel_replicas = optimization_settings.query_plan_with_parallel_replicas_builder();
+    if (!plan_with_parallel_replicas)
+        return;
+
+    const auto * final_node_in_replica_plan = findTopNodeOfReplicasPlan(plan_with_parallel_replicas->getRootNode());
+    if (!final_node_in_replica_plan)
+        return;
+    LOG_DEBUG(getLogger("optimizeTree"), "Top node of replicas plan: {}", final_node_in_replica_plan->step->getName());
+
+    const auto [corresponding_node_in_single_replica_plan, single_replica_plan_node_hash]
+        = findCorrespondingNodeInSingleNodePlan(*final_node_in_replica_plan, *plan_with_parallel_replicas->getRootNode(), root);
+    if (!corresponding_node_in_single_replica_plan)
+        return;
+
+    /// Now we need to identify the reading step that should be instrumented for statistics collection
+    ReadFromMergeTree * source_reading_step = findReadingStep(*corresponding_node_in_single_replica_plan);
+    if (!source_reading_step)
+        return;
+
+    const auto analysis
+        = source_reading_step->getAnalyzedResult() ? source_reading_step->getAnalyzedResult() : source_reading_step->selectRangesToRead();
+    if (!analysis)
+    {
+        LOG_DEBUG(getLogger("optimizeTree"), "Cannot get index analysis result from MergeTree table. Skipping optimization");
+        return;
+    }
+    const auto rows_to_read = analysis->selected_rows;
+    if (!rows_to_read)
+    {
+        LOG_DEBUG(getLogger("optimizeTree"), "Index analysis result doesn't contain selected rows. Skipping optimization");
+        return;
+    }
+
+    bool table_data_drifted_significantly = true;
+
+    const auto & stats_cache = getRuntimeDataflowStatisticsCache();
+    if (const auto stats = stats_cache.getStats(single_replica_plan_node_hash))
+    {
+        bool apply_plan_with_parallel_replicas = optimization_settings.automatic_parallel_replicas_mode != 2;
+        if (std::max<size_t>(stats->total_rows_to_read, rows_to_read) > std::min<size_t>(stats->total_rows_to_read, rows_to_read) * 2)
+        {
+            LOG_DEBUG(
+                getLogger("optimizeTree"),
+                "Significant difference in total rows from storage detected (previously {}, now {}). Recollecting statistics",
+                stats->total_rows_to_read,
+                rows_to_read);
+            apply_plan_with_parallel_replicas = false;
+        }
+        else
+        {
+            table_data_drifted_significantly = false;
+        }
+
+        if (apply_plan_with_parallel_replicas)
+        {
+            const auto max_threads = optimization_settings.max_threads;
+            const auto num_replicas = optimization_settings.max_parallel_replicas;
+            const auto local_plan_cost_estimation = stats->input_bytes / max_threads;
+            const auto replicas_plan_cost_estimation
+                = (stats->input_bytes / (max_threads * num_replicas)) + stats->output_bytes / num_replicas;
+            LOG_DEBUG(
+                getLogger("optimizeTree"),
+                "The applied formula: {} / {} ? ({} / ({} * {}) + {} / {}) ≡ {} ? {}",
+                stats->input_bytes,
+                max_threads,
+                stats->input_bytes,
+                max_threads,
+                num_replicas,
+                stats->output_bytes,
+                num_replicas,
+                local_plan_cost_estimation,
+                replicas_plan_cost_estimation);
+            if (local_plan_cost_estimation > replicas_plan_cost_estimation)
+            {
+                if (optimization_settings.automatic_parallel_replicas_min_bytes_per_replica
+                    && stats->input_bytes / num_replicas < optimization_settings.automatic_parallel_replicas_min_bytes_per_replica)
+                {
+                    LOG_DEBUG(
+                        getLogger("optimizeTree"),
+                        "Not enabling parallel replicas reading because {} < automatic_parallel_replicas_min_bytes_per_replica {}",
+                        stats->input_bytes / num_replicas,
+                        optimization_settings.automatic_parallel_replicas_min_bytes_per_replica);
+                    return;
+                }
+
+                ReadFromMergeTree * local_replica_plan_reading_step = findReadingStep(*final_node_in_replica_plan);
+                if (!local_replica_plan_reading_step)
+                    throw Exception(ErrorCodes::LOGICAL_ERROR, "Cannot find ReadFromMergeTree step in local parallel replicas plan");
+                chassert(local_replica_plan_reading_step->getAnalyzedResult() == nullptr);
+                local_replica_plan_reading_step->setAnalyzedResult(analysis);
+                moveSetsFromLocalPlanToReplicasPlan(query_plan, *plan_with_parallel_replicas);
+                query_plan.replaceNodeWithPlan(query_plan.getRootNode(), std::move(*plan_with_parallel_replicas));
+                return;
+            }
+        }
+    }
+    else
+    {
+        LOG_DEBUG(getLogger("optimizeTree"), "No stats found for hash {}", single_replica_plan_node_hash);
+    }
+
+    if (table_data_drifted_significantly
+        || optimization_settings.automatic_parallel_replicas_mode == 2 // automatic_parallel_replicas_mode == 2 enforces statistics recollection
+    )
+    {
+        auto updater = std::make_shared<RuntimeDataflowStatisticsCacheUpdater>(single_replica_plan_node_hash, rows_to_read);
+        source_reading_step->setRuntimeDataflowStatisticsCacheUpdater(updater);
+        corresponding_node_in_single_replica_plan->step->setRuntimeDataflowStatisticsCacheUpdater(updater);
+    }
+}
+
+}
+
+}
