+++ src/Processors/QueryPlan/Optimizations/considerEnablingParallelReplicas.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <Processors/QueryPlan/QueryPlan.h>
+#include <Processors/QueryPlan/Optimizations/QueryPlanOptimizationSettings.h>
+
+namespace DB
+{
+
+namespace QueryPlanOptimizations
+{
+
+/// Heuristic-based algorithm to decide whether to enable parallel replicas for the given query.
+void considerEnablingParallelReplicas(
+    const QueryPlanOptimizationSettings & optimization_settings, QueryPlan::Node & root, QueryPlan & query_plan);
+
+}
+
+}
